fix: fix login all #20

+++ quanlikhohang/deletenhaphanphoi.php
@@ -1,15 +1,14 @@
 <?php
-// session_start();
-// if (!isset($_SESSION['username']) || $_SESSION['role'] != 1) {
-//     header('location:1-displayAdmin.php');
-//     exit;
-// }
-
-
-
-
+session_start();
+if (!isset($_SESSION['user'])) {
+    header('Location: login.php');
+    exit;
+}
+require('../connection.php');
+?>
+<?php
 if (isset($_GET['deletedid'])) {
-    require 'connect.php';
+    require '../connection.php';
     $id = $_GET['deletedid'];
     $sql = "DELETE FROM nhaphanphoi WHERE id='$id'";
     if (mysqli_query($conn, $sql)) {
@@ -20,4 +19,3 @@ if (isset($_GET['deletedid'])) {
 } else {
     header('Location: dsnhaphanphoi.php?status=id_not_found');
 }
-?>
